another try to fix the installation problem -- moved files from madqc_v2 to chem folder
