+++ src/madness/chem/CMakeLists.txt
@@ -50,6 +50,15 @@ set(MADCHEM_HEADERS
     pointgroupsymmetry.h
     polynomial.h
     potentialmanager.h
+    ParameterManager.hpp
+    PathManager.hpp
+    MolresponseLib.hpp
+    Applications.hpp
+    Drivers.hpp
+    SCFTargetAdapter.hpp
+    InputWriter.hpp
+    CCLib.hpp
+    MoldftLib.hpp
     projector.h
     QCPropertyInterface.h
     SAP.h
@@ -87,6 +96,7 @@ set(MADCHEM_SOURCES
     mp2.cc
     nemo.cc
     NWChem.cc
+    InputWriter.cpp
     oep.cc
     pcm.cc
     pointgroupsymmetry.cc
