+++ src/madness/chem/InputWriter.cpp
