+++ src/madness/chem/PathManager.hpp
